[FIX] iot: download Worldline lib as odoo user
As of iot box image >= 25_01 we introduces a user "odoo" which is
responsible for everything related to the Odoo service on the IoT Box.

However, the Worldline related files are being downloaded and unzipped
under 'sudo' command which means that the user "odoo" doesn't have
access to them

This is an issue when trying to delete the old IoT handlers as it's done
with the user "odoo" but Worldline files presend in the 'ctep' directory
belong to 'root'

This PR runs Worldline related script under user "odoo" which fixes the
ownership/access issues and allows Odoo service to delete the files.

task-4527243

closes odoo/enterprise#78611

X-original-commit: 9333148953d43b3de00b0d868393dc6cb2385938
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>