[FIX] web_studio : Remove redundant code
Issue:
------
When making changes in version 17.0 using studio the correspondent standard
view's noupdate becomes 'True' as per the logic implemented. But when we migrate
the database to the further versions like 18.0. The changes made in that
particular view in 18.0 will not be loaded and will cause errors since its
inherited view's will not be compatible enough.

For ex:
If we make changes in a "Purchase Order"(i.e; purchase.report_purchaseorder)
report then after saving the changes the noupdate of that report view
which is "purchase.report_purchaseorder_document" will be updated to 'true'.
And the action record i.e; "purchase.action_report_purchase_order"
 noupdate will be updated to 'true'.

 https://www.awesomescreenshot.com/image/52484998?key=f9c3a20e7a50a43b346e37836cea21c4
 https://www.awesomescreenshot.com/image/52484997?key=eeb3208738a5891ac8c507b45fab50dc

If the customer makes changes using studio then those changes will be recorded
in the studio view and there is no point of making the standard view's noupdate
to 'true'.

Solution:
-----------
Remove redundant code to overcome updating noupdate to 'true'.

Steps to reproduce:
------------------------
1. Create a database in version 17.0.
2. Make changes in any report using studio and save.
3. Migrate the database and try opening that report.
4. You'll encounter issues due to incompatibilty of views.
    Since, the changes made in the standard will not be loaded
    because it's noupdate is true.

REF PRS:
Enterprise PR:
1.https://github.com/odoo/enterprise/commit/2ae63d9a43ba973d905232105201d88e423a8245#diff-cf8f46a54f86279b4d773f0d39b23d22f046eb6d33a39c851f2a86dcd2b5b202R463

closes odoo/enterprise#78582

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>