[FIX] documents_spreadsheet: tours that don't properly wait for unload
During testing of the "new" Chrome's headless mode (by default since
version 128), a concurrency issue arise at the end of those tours: a
last request to the server is actually sent during the unload of the
view. This request actually reach the server even if Chrome is already
closed and the test is cleaning itself (rollbacking the changes made in
the database), resulting into a "release savepoint" mismatch on an
already aborted transaction (example build's error [1]).

This commit is kind of a follow-up of a previous PR [2] where a last
step allowing for the spreadsheet to unload was added and adding one
more to actually wait to be back on the Document app.

[1] https://runbot.odoo.com/runbot/build/74602509
[2] https://github.com/odoo/enterprise/pull/51795

closes odoo/enterprise#78797

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>