[FIX] account_online_synchronization: Avoid using recordset to call model function
This commit avoids calling the toggle cron function with
journal_ids recordset after an online account has been unlinked.
Previously, this caused an error because the record was already
deleted. The journal_ids model is now retrieved from the environment
instead of the potentially stale recordset.
The bug was introduced by this commit [[1]].

no task id

[1]: https://github.com/odoo/enterprise/commit/bc3b9136d4d62fbc1d09267f73cac7ca29f1ba3a

closes odoo/enterprise#78744

X-original-commit: 4a5b69140fd7b76e83e27a4f977d00b9752d9f31
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>