[FIX] account: prevent traceback for transaction_details with control characters
In the bank reconciliation widget, when the transaction_details of bank
statement lines get formatted by calling json.loads(), an error will show if
some strings contain control characters (\n, \t, \r, \0).

Adding the paramter `strict=False` to the call to json.loads would allow
such characters to be kept by the formatting.

In the view, they will not be used (any of those 4 chars will simply be
considered a whitespace).
One could add `style="white-space: pre-line"` as an attribute to the span tag
of those element, but this is not necessary. This PR simply aims to patch the
already existing data in databases.

opw-4480250

closes odoo/enterprise#78684

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>