[FIX] stock_barcode{_mrp}: correctly extract kanban tip messages
The kanban tip messages displayed at the top in the Barcode app were
constructed with `<b>` markup in the OWL template. However, the
translation extraction mechanism for OWL templates splits the template
up per tag it encounters. So if you had a message like
`Scan a <b>transfer</b>, a <b>product</b> ...`, it would extract the
`, a` part separately.

Moreover, if the final string didn't contain more than 1 alphanumerical
character, it wouldn't be exported in the .pot file and thus not be
translatable.

Also, having these messages chopped up doesn't allow translators to
change word order in other languages.

In this commit we safely construct the messages in Javascript and have
them translated as one full sentence. After this, we show them in the
OWL template.

task-4421055

closes odoo/enterprise#78717

Related: odoo/odoo#196805
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>