[FIX] mrp_account_enterprise: correctly compute mo_qty with multiple MOs
Bug introduced by this commit: https://github.com/odoo/enterprise/commit/4b312d56e7fa01ab4e8f534a1e62ef59e1421f77

Steps to reproduce the bug:
- Create a storable product “P1”
- Create two different MOs:
    - First MO: 1 unit of P1
    - Second MO: 2 units of P1
- Confirm and mark them as done
- Go to the manufacturing list view
- Select both MOs
- Print Cost Analysis

Problem:
The mo_qty is 2 units instead of 3;
only the quantity of the last MO is taken into account. This is because,
to calculate the total quantity of all MOs, we use a variable mo_qty,
but it is initialized twice, once before the loop and again within the MO
loop, so for each MO, the counter is reset to 0.

opw-4529333

closes odoo/enterprise#78731

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>