[FIX] account_accountant: save when removing analytic rec line
How to reproduce the issue:

- Create an analytic distribution model.

- In the bank reconciliation, create a new transaction. In manual operations, remove the analytic.

If you reload the reconciliation, the analytic distribution is not deleted.

In the `bank_rec_widget` of account_accountant, the method `_line_value_changed_analytic_distribution` is triggered when the analytic field is modified. However it does not save the new value of the analytic on the account.move.line.

opw-4460079

closes odoo/enterprise#81504

X-original-commit: 8f30a651aefd13293eb0cdeaf2fb73d3544d8552
Signed-off-by: John Laterre (jol) <jol@odoo.com>