[FIX] frontdesk: keep "Welcome to" together for translations
Currently on the Frontdesk welcome screen the words "Welcome" and "to"
(welcoming visitors to your company) were exported separately, because
they weren't in the same parent node. This caused translation issues,
since the word "to" is translated differently depending on the context
and was not scoped to the module (Javascript translation).

This commit puts "Welcome to" together in the same parent node, so the
phrase is exported as a whole.

task-4421055

closes odoo/enterprise#81898

X-original-commit: d503c3be2da377989e0be960f143eec1c08c5cf0
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>