[FIX] frontdesk: keep "Welcome to" together for translations
Currently on the Frontdesk welcome screen the words "Welcome" and "to"
(welcoming visitors to your company) were exported separately, because
they weren't in the same parent node. This caused translation issues,
since the word "to" is translated differently depending on the context
and was not scoped to the module (Javascript translation).

This commit puts "Welcome to" together in the same parent node, so the
phrase is exported as a whole.

task-4421055

closes odoo/enterprise#78781

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>