[IMP] account_report: caret builder on groupby keys
If a custom handler handles the caret builder, it used to only consider this function,
and therefore not benefiting from the generic keys provided by the default function of
the report engine. This can cause minor disadvantage such as when a user change himself
the groupby keys of a report with a custom engine. In this case, the caret options won't
be visible at all. With the change of this commit, the custom handler caret options will
be used on top of the default ones.

task-4557882

closes odoo/enterprise#78800

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>