[FIX] mrp_workorder: by_pass_set_qty_producing for comp in operations
Steps to reproduce:

1. Create a bom for a prodcut FP with one tracked component, one
   untracked component, and one operation
2. For each component line, specify that they are to be consumed in that
   operation ('consumed in operation' column)
3. Create a manufacturing order with that bom and confirm it
4. Set the quantity producing of the MO to 1

Issue:

A reservation is made for the tracked component.

Cause of the issue:

Setting the qty_producing on the MO will reserve the component move raws
as soon as they are not flagged as _should_bypass_set_qty_producing:
https://github.com/odoo/odoo/blob/9c3e872137841c13dec89bcfe4f83bed9f0e7efa/addons/mrp/models/mrp_production.py#L1218-L1228
However, this method has become a complex gaz powerplant:
https://github.com/odoo/enterprise/blob/f5c5a4a4c85adcd5bc5d9e996ed2114d4d9cb802/mrp_workorder/models/stock_move.py#L10-L16
And in this basic use case, even thought the component is expected to be
registered in the operation, it will not trigger the baypass.

Fix:

We backport the definition of the filtering method as it was in 17.2:
https://github.com/odoo/enterprise/commit/536886ea663a66fc79fb9bea5e62280228802d8f

opw-4442455

closes odoo/enterprise#78973

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>