[FIX] sale_amazon: upgrade SP-API inventory feed
Amazon SP-API is dropping support for the
`POST_INVENTORY_AVAILABILITY_DATA` feed type. This was previously used
in Odoo to synchronize the stock level with Amazon.

This commit upgrades the feed api to use the new `JSON_LISTINGS_FEED`
as a replacement.

This new feed needs new informations that were not stored previously.
Since this is a fix targeting stable versions of Odoo, it was decided to
store this new inforamtion as a json dictionary in the
`amazon.offer.amazon_feed_ref` field.

task-3820591

closes odoo/enterprise#78897

Signed-off-by: Lionel Piraux (lipi) <lipi@odoo.com>