[FIX] sign: add necessary attributes as query parameter
- Sign > Documents > Click on any of the Kanban record
- Reload the view (either reload the browser, or activate
the debug, or change to dark mode on the user menu).

Before this commit, on reloading the sign document page user was
redirected back to kanban view. This occurs because, the client
action 'sign.Document' requires  attributes which are found in the
context - { id, token, create_uid, state, request_item_states} that was
lost when reloading.

Now, { id, token, create_uid, state, request_item_states} is put in the
query string of the URL, in that way, when reloading, the client action
'sign.Document' will have the needed attributes. Note that, this is also
the behavior of the base_import 'ImportAction' action [1].

task-4391729

[1] : https://github.com/odoo/odoo/commit/a6801ce4aa65cd023e2a762ab5c42755a4599c95

closes odoo/enterprise#80775

X-original-commit: 2381354c8846ab9076e0f99d20cc8796dde13327
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Saurabh Choraria (sauc) <sauc@odoo.com>