[FIX] sendcloud_service: Send average weight for multipackage shipments
Before this commit:

Total weight sent for a multipackage shipment was sum of all the weights of packages in the shipment.
Which causes shipping product with wrong weight limits to get selected.

After this commit:

Average weight of all the packages is sent as total weight.

Sendcloud in the backend use average weight for the multipackage shipping request and
multiplies it by the number of packages. *checked with sendcloud support.

opw-4260104

closes odoo/enterprise#79001

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>