[FIX] account_reports: avoid double call to _custom_options_initializer in case of manual variant creation of a custom report
To reproduce:
(Note that this needs to be donne before https://github.com/odoo/enterprise/commit/03fd8fa9ce59f3b052ba112612ee288cd44f045a ; you can easily replicate by indenting the add of the "Send" button one more level to the left, so that the button is always added, as it used to be)

1) Duplicate the Partner Ledger
2) Set the original Partner Ledger as the root report of the duplicate
3) Open the Partner Ledger, and select the duplicate in the variant selector

===> The "Send" button is displayed twice on top of the report.

This is due to the fact the custom handler is first called by the report itself (since the duplicate kept the same handler as the root Partner Ledger), then by the call to super() made in the _custom_options_initializer.

In the case of the Partner Ledger, we could have removed that super call (since it's a root report), but we prefer a more generic and cleaner fix that just makes sure we don't reinitialize the same custom options if we notice the handler is the same.

opw-4506968

closes odoo/enterprise#79101

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>