[FIX] account_accountant: Invoicing switch threshold UI fix
Change the "up to this date" to "before this date", to specify that its
upper limit is excluded.

Otherwise, when you set 31/12/24 in this field and Odoo cancels all the
entries up to 30/12/24 and leave the 31/12/24 entries, you'll be
disappointed.

Task [link](https://www.odoo.com/odoo/project/967/tasks/4447369)
task-4447369

closes odoo/enterprise#79505

X-original-commit: 1fc7a34adb8848bd90780fb306c990c0d5bb5e6d
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>