[FIX] knowledge: fix embedded calendar quirks
In Knowledge, in the Calendar embed, event text descriptions
and properties tend to overlap with each other, and the header.
The Calendar embed also does not resize properly when the
sidebar width or article width get changed.

This commit makes it so that event descriptions and properties
do not overlap ; and that calendar width (and, subsequently,
column width) is affected by sidebar or article width
changes.

task-4558391

closes odoo/enterprise#80477

X-original-commit: 591972d8ce4f9f51ee68eb25137b3f9a4fa07692
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>