[FIX] knowledge: fix various UI bugs/janks
In Knowledge, on chromium-based browsers, zooming at certain
levels will cause the comment icon to flail around the editor
and lock in unwanted places. (This behavior was not reproductible
in Firefox.)

To circumvent this, the margin calculations for the purposes of
calculating the comment icon horizontal location are now done
manually.

task-4558391

closes odoo/enterprise#79165

Related: odoo/odoo#197468
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>