[FIX] account_asset: cannot resume asset with gain
An error is encountered in the Modify Asset wizard when resuming Asset Depreciation
after manually increasing Non Depreciable Amount. The system attempts to record
the change as a gain, requiring a Depreciation Account. However, this field is
hidden, preventing users from fixing the problem, so they are blocked.

Steps to reproduce:
- Create a running asset
- Click Modify Depreciation > Pause
- Click Resume Depreciation (Asset modify wizard will open).
- Increase Not Depreciable Amount, realizing a gain
- Press the Resume button

Issue: Error will raise

opw-4472309

closes odoo/enterprise#82512

X-original-commit: 56b333080c08252a53f5d28ca89ae902e44901fa
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>