[PERF] account_accountant: batch _generate_deferred_expenses
Before this commit, the function _generate_deferred_entries was
basically a big for-loop in which a lot of account.moves and lines are
created in each iteration. This was extremely inefficient because it
wasn't leveraging any batch optimization of the ORM. This PR breaks down
the for-loop into smaller ones that aggregate vals and performs a single
call to create records, instead of individual creations. Same approach
was taken for unlinked records and posting moves.

The assumption is that records are created in the same order of the
values list. This allows for breaking the for-loop because consecutive
for-loops can simply inherit the order from the lists created in the
loop just before. Another leverage point is that simply adding move_id
to account.move.line(s) before creation is equivalent to
move.write({'line_ids': \[...\]}). Based on the 2 assumptions, we could
decouple the creation of moves and their lines into separate loops to
aggregate their values.

After this PR creations, unlinks and posting operations are all batched
to leverage the power of ORM optimization.

Benchmarks:
|Num deferred expenses | Time before | Time After | Num queries before | Num queries After |
|--------------------- | ----------- | ---------- | ------------------ | ----------------- |
| 100 		| 23.9s 	| 13.1s 	| 3768 		| 2918 			   |
| 1,000 	|181.4s      	| 66.5s       	| 17885         | 9530              	   |
| 10,000	|TIMEOUT(>1200s)| 616,7s    	| 1379          | 1010              	   |

opw-4480919
opw-4403217

closes odoo/enterprise#79255

Signed-off-by: John Laterre (jol) <jol@odoo.com>