[FIX] industry_fsm: disambiguate report query
The project report query joins multiple tables and
aggregates on some fields, everything else should be
in the group by clause. The group by here is using the
select alias but it is better to use the full column
name including the table alias, in case one of the
joined tables has a column with the same name as the
alias.

closes odoo/enterprise#79326

X-original-commit: 76b6f17b66e668ec09e01c29e929ee1d21ebec73
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>