[FIX] spreadsheet_edition: error when renaming version
Clicking 'Name this version' in the version history panel causes a
traceback because the `el` of TextInput is not found.

We have two possible solutions:
1. Forward the refs of TextInput and VersionHistoryItem explicitly
using `useChildRef` and `useForwardRefToParent`, or use a callback to
retrieve the ref for focusing the input.

2. Remove the menu item, as the cursor already changes to a text
pointer on hover, indicating the name is editable.

I think the second approach is better since it is simpler and
more intuitive.

closes odoo/enterprise#79474

Task: 4563422
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>