[FIX] sale_subscription: apply good email template on subscription
**Problem:**
When reversing an invoice and making a credit note on a subscription, the email template was still "Invoice: Sending" instead of choosing the credit note template "Credit Note: Sending".
The condition on the following code was always True, meaning the super method, that is used to decide which template to take in case of a credit note, was never called.
https://github.com/odoo/enterprise/blob/eb1eff85d2119a34997067fedb602e963a5d6f02/sale_subscription/models/account_move_send.py#L9-L14
**Steps to reproduce:**
- Take an invoice created through a subscription, reverse it (credit note action)
- Try to send it
- See that the email template is "Invoice: Sending" and not "Credit Note: Sending"
**Fix:**
A condition was added to differentiate an invoice from a credit note.

opw-4455155

closes odoo/enterprise#79331

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>