[FIX] stock_barcode,*: batch grouped lines
Description of the issue
========================

The initial issue was, when tracked product lines are grouped by product
in batch ("Group batch lines" parameter checked in the picking type
config), the sublines weren't displayed.
This issue was caused by the way we choose to display or not
sublines. For a tracked product, we don't display them if reserved
lot/serial is hidden (picking type config) and if there is no qty done.
That said, that didn't take in consideration the grouping by product
where lines for different pickings are grouped together.
In this case, it's better to show the qty demand anyway to know how many
quantity needs to be processed for each picking.

Once this issue is fixed, a second issue appears.
Imagine you have this following case:
- 1 line for lot product (0/4 qty) for receipt 1;
- 1 line for lot product (0/4 qty) for receipt 2.
- In batch, these lines are grouped because they share the same product.

Now, let's say we do the following process:
1. Scan the lot product -> The grouped line is selected;
2. Scan lot01 two times -> receipt 1 line is now 2/4 qty with lot01;
3. Scan lot02 -> receipt 2 line is now 1/4 with lot02.

The point 3. where is the issue happens. Usually, when we scan a lot, we
check if the line is completed or not by checking its parent line's
quantity. Here, the parent line is the batch grouped line and should not
work exactly like that, because by checking this line as the parent of
every sublines means we can't know if a particular picking need is
fulfilled or not.

How to reproduce
================

- Inventory > Configuration > Operations Types > Barcode App:
  - Uncheck "Show reserved lot/SN";
  - Check "Group batch lines".
- Create two receipts for the same tracked by lot product;
- Create a batch for this receipt;
- Open the batch in the Barcode app -> You can see there is a line for
  this product with the total quantity but when clicking on the button
  to display sublines, the sublines are not shown.

Fix
===

For the first issue, we simply added a condition for
`stock.picking.batch` to adapt to this case.

For the second issue, the group by tracked product inside a single
picking and the group by product inside a batch are not the same and use
a different key.
The first kind of group uses the key `parentLine` and the second one use
the key `batchParentLine`. When we want the parent line of a subline, we
will always refer to the `parentLine`. But in the interface, if this
parent line is itself inside a `batchParentLine`, the `parentLine` won't
be displayed, only the batch parent line.
That way, the user still has all lines for a same product grouped
together under a single parent line, but to know if a need if fulfilled,
we still check on the parent line limited to a single picking.

opw-4291149

closes odoo/enterprise#79516

Signed-off-by: William Henrotin (whe) <whe@odoo.com>