[FIX] account_followup: fix extra spaces in chatter and pdf
In the follow up process, while the wizard's text is displayed correctly, the text in the chatter and in the pdf would contain too many carriage returns.

The error is linked to the use of the text widget in the xml, which doesn't seem necessary in this case and is therefore removed.

task-4574941

closes odoo/enterprise#79925

X-original-commit: 8af89ac742c1c627914663b0c6b9e84015318a81
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Camille Spiritus (casp) <casp@odoo.com>