[FIX] l10n_cz_reports_2025: manifest was preventing to auto install
Because of the countries flag set in the module manifest, the module was set to be auto-installed only when there was a company with country 'CZ' existing. Although this can't really be pointed as a mistake, it was causing an error in account_reports TestAllReportsGeneration.test_generate_all_export_files, as we'd be simply installing the module l10n_cz_reports and expect everything to be set and tested.

runbot error: 114355

closes odoo/enterprise#80080

X-original-commit: e11240374a157ed6641ebeb84ca0260d0f54c6d5
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>