[FIX] *_extract: fix IAP account used during validation
During the execution of the validation cron of the OCR, the IAP account
selected could sometimes be incorrect.

In the context of the validation cron, we need to know to which company
the document belong in order to get the correct IAP account (i.e. the
one from the company, if it exists).

If we use the generic IAP account (i.e. the one that isn't linked to any
company), the mismatch of accounts will cause the validation to be
ignored by the OCR server.

Related ticket #4522628

closes odoo/enterprise#79675

X-original-commit: ead0d59be52a841a96cdd0592281d326348325be
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>