[FIX] planning, planning_holidays: fix flexible resource leave in planning gantt
Prior to this commit, flexible resources did not have their leaves
reflected as gray cells in the planning gantt view. This was due to the
work_intervals being ignored for the calculation of flexible resources
availability.

This commit adds measures to handle the leaves for flexible resources by
setting a dummy attendance (which covers the whole length of the period
in gantt interval), and then injects their leaves intervals. Please note
that this is aimed for flexible resources only, and not for fully flexible
resources as these are not considered to take leaves.

To replicate:
1. set timeoffs to a flexible resource (whole day and half day)
2. open planning app
3. in the gantt view, the day in which the timeoff was set for a whole day
   should be fully grayed (in the default month granularity view)
4. when selecting only a day as the granularity of the gantt view, the day
   in which the timeoff was set for a half day should be half grayed.
   (from 00:00-12:00 in case of a morning timeoff, and from 12:00-24:00
   in case of an afternoon timeoff)

ticket-id: 4492625
community:

Part-of: odoo/enterprise#79532
Related: odoo/odoo#198032
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>