[FIX] account_winbooks_import: avoid computing the payment_state
Since all moves are imported as draft, it doesn't make sense to try to
compute the payment state.
Moreover, when importing a lot of lines in a fresh database, the stats
of the table are completely off. This leads to potentially a very bad
plan made when fetching all the data needed to compute the field.

Part-of: odoo/enterprise#79533
Signed-off-by: John Laterre (jol) <jol@odoo.com>