[FIX] stock_barcode{,_mrp}: avoid _truncate_overreserved_moves on done moves
Issue 1:

- Create a storable product with a BoM:
    - 1 x Comp as component
- Create and confirm a new MO for 1 unit of that product
- Open barcode, uncheck the ready filter > select the MO
- Click Produce without scanning anything.
> Traceback

Cause of the issue:

Clicking on produce will first validate the MO, setting the desired
qties on the moves and picking them and then trigger a `'history-back'`
on the previous tab to leave the record:
https://github.com/odoo/enterprise/blob/ead0d59be52a841a96cdd0592281d326348325be/stock_barcode/static/src/models/barcode_model.js#L475-L495
As such, the component will be destroyed and a "post_barcode_process"
will be launched:
https://github.com/odoo/enterprise/blob/ead0d59be52a841a96cdd0592281d326348325be/stock_barcode/static/src/models/barcode_picking_model.js#L1354
Unfortunatelly, the `_truncate_overreserved_moves` part of this
`post_process` was not design to work with already validated moves.

Issue 2:

The float compare used in the `_truncate_overreserved_moves`
sets a rounding as precision digits rather than precision rounding.

opw-4353884 and opw-4417870

closes odoo/enterprise#79653

Signed-off-by: Steve Van Essche <svs@odoo.com>