[IMP] pos_restaurant_appointment: Add minimal html editor in pos mode
In the point of sale environnement we want to be able to edit html field
in a minimal way. So only core plugins are added to the html_editor.

As we have a special assets environnement for the point_of_sale we can
directly patch the main `htmlField` definition without impacting main
assets of Odoo.

closes odoo/enterprise#79703

Community: 198482
Related: odoo/odoo#198482
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>