[REF] documents: add basic documents view hoot tests
Also impacted: documents_spreadsheet tests.
As the owner of documents is required to appropriately assign the location
of root documents in the app, odoobot is made the default owner (so documents
are created in Company).

This required to make sure that "res.users" records used (odoobot) exist, which can
be done with a minimal-ish diff versus the alternatives, and is future-proof for
other Documents-related requirements.

Task-3861500

Part-of: odoo/enterprise#76201
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>