[FIX] *: misused manual savepoint
The `rollback` decorator of `account_acountant` does not properly
`release` its savepoint, leading to the subtransaction remaining
present until the end of the current transaction (or a rollback to the
start of the test).

This is unlikely to trigger odoo/odoo#71395 (though not impossible as
one of the savepoints is in a test) but it's still bad form and
unnecessary: an explicit `close` will *rollback and release* the
savepoint, so `savepoin()` can just be wrapped into a `closing()`
which will do that for us.

For `web_gantt` the savepoint is closed but not on all paths (if
`ValueError` is raised the savepoint is never rolled back let alone
closed), the manual use is also rather unnecessary as we can just
rollback around the relevant method call, none of the later mess
interacts with the cursor / connection.

closes odoo/enterprise#79837

X-original-commit: 6b4bc664ed4e8112d60cd6479f7c97f6b479ee92
Related: odoo/odoo#198568
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>