[FIX] social_twitter: reenable api calls to X
This commit fixes an issue with the X api calls where the oauth_signature was no
longer correct in recent Python versions.

`url_quote` used to consider the character `,` as unsafe and quoted it when
creating the signature. From Python3.11 this method consider it as reserved but
not as unsafe thus not quoting it. This change restores the old behavior to
quote so as to respect the oauth_signature format given by X's doc.

task-4583612

closes odoo/enterprise#79871

X-original-commit: d3a84c0390a03d637ce89ffda5c5b6d8483defbe
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>