[FIX] l10n_co_edi_website_sale: revert modifying website_sale template
In 954e2110c we modified the `website_sale` template when
l10n_co_edi_website_sale is installed, to always show the VAT fields.

This modification was applied to all websites (regardless of whether in
Colombia or not) and was bypassing the code that checks whether the VAT
field can safely be modified or not from the e-commerce.

We revert it and instead, if possible, will solve the original issue
(VAT fields not available to be edited for billing addresses) by
redirecting the user to the parent address to edit the VAT.

opw-4309448

closes odoo/enterprise#79882

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>