[FIX] sale_commission: display good aggregation and access pivot
**Problem:**
This is a reverse and improvement of a previous fix. The first PR is this one : https://github.com/odoo/enterprise/pull/78537
This fix caused the pivot view to be inaccessible, creating a traceback, because the field 'achieved_rate' needs to be able to be aggregated. The line that removed the aggregator that was added was the new problem.
https://github.com/odoo/enterprise/blob/137b233d4990ea9f48c8aa5b6a3990dc4170356b/sale_commission/report/commission_report.py#L19

**Steps to reproduce:**
- Go to commissions in the sales app
- Click on the pivot view
- The traceback will appear

**Fix:**
Change the aggregator from None to 'avg' to pick the average of the achieved_rate in case of a group by.

opw-4512739

closes odoo/enterprise#79912

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>