[FIX] project_enterprise: show gantt without stages enabled
**Issue**
The project's Gantt view doesn't work if the "Project Stages"
setting is not enabled.

**Steps to reproduce**
- Install `project_enterprise` and Studio.
- Have project stages disabled.
- Open Projects app.
- Click on Studio button > Views > Try to activate Gantt view.
-> `AccessError for read on stage_id`

**Cause**
The `stage_id` field used on the project gantt view is only readable if
the project stages feature is enabled.
https://github.com/odoo/enterprise/blob/b911ed1e857e1534c86495cd4934c1adb8ae9597/project_enterprise/views/project_views.xml#L12

**Solution**
A hack is to add in the model metadata the `colorField` key used by the
gantt renderer to color the pills:
https://github.com/odoo/enterprise/blob/28f2fba988b1b225c97d15729020962ed4dee73b/web_gantt/static/src/gantt_renderer.js#L778
It is normally set when parsing the view here:
https://github.com/odoo/enterprise/blob/dfe2795b820527a8226d8cf1afa5bf7ac8fcb53d/web_gantt/static/src/gantt_arch_parser.js#L194
This avoids trying to read the field if the user doesn't have the
necessary `project.group_project_stages` group.

opw-4592110

closes odoo/enterprise#81388

X-original-commit: 7720345619564da8ad8c3c1611a24cefbb2c0d61
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Tanguy Quéguineur (taqu) <taqu@odoo.com>