[FIX] l10n_fr_reports, account_reports: restore reports from 2024
Issue:
The COA for the French localization has been updated in 2025.
Some accounts have been replaced or removed and the reports
have been updated with the new formulas.
However, the reports from 2024 are still needed for those who
need to print a report for 2024.

Solution:
Restore the old reports from 2024 as variants.

opw-4566689
opw-4590293
opw-4572865

closes odoo/enterprise#80203

X-original-commit: 4e6214473831bddf9167671a40376df8c251b3b9
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>