[FIX] project_enterprise: fix test that depends on timesheet modules
Before this commit, the `test_smart_schedule_with_allocated_hours_and_deadlines` test
fails when only `project_enterprise` module is installed because the
allocated_hours on the tasks are recomputed. However, when
timesheet_grid module is installed, the project inside the test has the
timesheets feature enabled by default and so the allocated_hours are not
recomputed because we suppose the allocated_hours on tasks inside that
project can only be altered by the user when the tasks are already
created.

This commit checks in the test is timesheet_grid module is installed to
give the right expected allocated_hours on tasks.

runbot-108336

closes odoo/enterprise#81387

X-original-commit: 2b675fd8d299ef0f4a90b47c773b9dca92a4ca59
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>