[FIX] project_account_budget: change color of expense budget
How to reproduce the issue:

- Create an expense budget and associate it to a project.
- Create invoice/bills related to the project

In the project dashboard view, the remaining value will be marked in green for expenses that exceed the budget (over-achieving costs) and in red for expenses that are under the budget (under-achieving costs). However, this was deemed non-intuitive, as under-achieving costs should be seen as a positive indicator.

After this commit the color for expense budget are inverted.

opw-4562815

closes odoo/enterprise#79965

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>