[FIX] documents: check if element exits before calling `scrollIntoView`
Issue:
- When we select a document and click on the "Info and tag" button,
a traceback is thrown.

Steps to reproduce:
- Open the documents app
- Preview a document
- Click on the "Info and tag" in the "Action" dropdown menu
- a traceback is thrown

Solution:
- When we click on the "Info and tag" button, `onToggleChatter` is
called, in it we try getting the element of the selected document
and scrolling to it, if the `querySelector` returns null, and we try
to call `scrollIntoView` on it, a traceback is thrown.
- we fixed this by adding the optional chaining operator `(?.)` to
safely call scrollIntoView only if querySelector returns an element.
- although not directly related to the issue, we also added a small
check to disconnect any previous `MutationObserver` before creating
a new one.

OPW-4586089

closes odoo/enterprise#79975

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>