[FIX] web_studio: field option boolean number
Have a field in a view like this
```<field name="" options="{ some_boolean: 1 }" />```

And have the corresponding supportedOption (defined on the field widget):
```
{
	somme_boolean: {
		type: "boolean"
	}
}
```

Open studio and click on that field, in debug mode.

Before this commit, there was a crash because 1 or 0 were directly passed to a component which expected a boolean value.
After this commit, before displaying the options we convert 0 or 1 to the corresponding boolean.

closes odoo/enterprise#80014

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>