[FIX] web_studio: eval conditional t-set
In the external and internal layouts, some default variables are set conditionally (`o` in particular)
This commit evaluates the content of the t-if expression to imitate what the real qweb does.
Otherwise, the variable risks of being erased from the context of the node, which is necessary to
add field on the right variable with the right type.

opw-4592446

closes odoo/enterprise#80019

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>