[FIX] documents: set `allowed_company_ids` on upload
Issue:
- When we user-defined defaults based on company_id set on documents,
they always trigger based on the user's default company instead of
the current company.

Steps to reproduce:
- In a multi-company environment say we have companies A(id: 1) and
B(id: 2), and the user default company is A. (he has access to both).
- Create 2 user-defined defaults for the field :
`documents.document.company_id`, setting value to 1 for company A and
2 for company B.
- Create a document using company B, the document's `company_id` will
be set to company B which is wrong.

Solution:
- in controllers `request.env.company` always returns the user's
default company, so whenever we upload a document (call the
`documents_upload` controller), `documents.document` create assumes that
the current company is the user's default company.
- To fix this, we add the `allowed_company_ids` to the upload request
params in the JS `DocumentService` and  we intercept in the python
controller and set the `allowed_company_ids` context on the request
to make sure create will use the current company instead.

OPW-4507424

closes odoo/enterprise#80060

Signed-off-by: Florian Charlier (flch) <flch@odoo.com>