[FIX] {account,l10n_be}_reports: enable export report buttons with branches
To replicate:
1. Install a localization reports module (e.g. l10n_lu_reports)
2. Create two branches for the company in the localization: one with
a different VAT, and another without VAT (it takes the same VAT as
the closest parent with VAT)
3. In Accounting > Reporting > Tax Report
4. Select only the company with a different VAT. The button to export
an XML is not enabled.
5. Select the parent company and the company without VAT. The button
to export an XML is not enabled.

Issue:
In some Tax Reports and EC Sales Reports with branch companies, the
additional buttons are not enabled properly, even when all companies
with the same VAT are selected.

Expected behavior:
If all branches with the same VAT are selected, export buttons should
be available.

Cause of the problem:
In the tax report handler and EC sales report handler,
`_custom_options_initializer()` calls the function
`_enable_export_buttons_for_common_vat_groups_in_branches()`, which
should enable the buttons if branches with the same VAT are selected.

When extending `_custom_options_initializer()`, the call to super
happens before extra buttons are added. These additional buttons are
not included in the call to
`_enable_export_buttons_for_common_vat_groups_in_branches()`.

Fix:
Instead of calling the function in _custom_options_initializer, add an
option key to check and enable export buttons. This check is now done
in `get_options()`, using the same logic previously applied in
`_enable_export_buttons_for_common_vat_groups_in_branches`. As
`get_options()` is called after the custom options initializer, all
buttons are enabled if all branches with the same VAT are selected.

task-4061157

closes odoo/enterprise#79227

Signed-off-by: John Laterre (jol) <jol@odoo.com>