[FIX] appointment: fix missing dependency in website
The appointment addon adds the `@mail/utils/common/format` module
to the frontent assets. However, this module depends on another
one: `@mail/utils/common/html`, which isn't there. As a consequence
there's a red banner displayed in the bottom left corner,
indicating that there's an odoo module issue, in all websites when
the appointment addon is installed. This is causing a lot of
trouble with free trials.

This commit adds the missing dependency.

closes odoo/enterprise#80091

X-original-commit: d4e5db539b736750c5c5accd143d590f87d84c54
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>