[FIX] sign: modal error on tour
Before this commit, the runbot would fail with Script Timeout when executing
the shared_sign_request_tour tour. The main problem was the title of the modal
being changed from "All done!" to "It's signed!" without updating the tour
itself. The secondary problem was the limitation added in the tour of clicking
a button inside a modal: this one is not allowed anymore due to a change in tours.

After this commit, we removed the 'Download button' trigger in the tour as it
is not allowed anymore and updated the string for the correct modal title in
the tour, from "All done!" to "It's signed!".

runbot-98514

closes odoo/enterprise#80070

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>