[FIX] pos_preparation_display: ignore inactive linked configs on load data
**Problem**:
When archiving the only pos config linked to the preparation display, the configuration of the preparation display shows that no pos config is linked to it (it's linked to all pos configs).
However, when pos config starts, it doesn't load the preparation display on _load_model as the domain for it is to load when the specific pos config is linked, or no configs are linked.
This confuses the clients as the archived pos will not be seen in the preparation display configuration.

Additionally, there is another issue related to live synchronization in version 18.0. When clicking the "Order" button, the request is sent only if the file is linked to a preparation display
see here https://github.com/odoo/enterprise/blob/18.0/pos_preparation_display/static/src/override/point_of_sale/pos_store.js#L37.

As a result, if the display is linked to an archived POS config, live synchronization will not function correctly.

- How to reproduce:
    * Assign one pos config to preparation display
    * Archive the assigned config
    * Open another pos config.
    * Kitchen display is not loaded

**Solution**:
Change the domain to include displays with no configs, or with no active config.

opw-4504623

closes odoo/enterprise#80117

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>