[FIX] l10n_mx_reports: filter account lines correctly
Currently, an error may occur when we try to export the trial balance
with the SAT (XML) format. This is because the account_group ID is used
to browse the account, due to a more specific account_group.

Steps to reproduce:
1. Install l10n_mx_reports.
2. Create an account_group with a more specific account, such as
401.01.01
   - The 4, 401, and 401.01 groups are created with demo data.
3. Go to trial balance and try exporting with "SAT (XML)".
4. An error will occur (Missing Record) because the account_group ID is
used to browse the account.

This fix does two things:
1. It ensures that only account.account lines are considered.
2. It changes the line length condition from `= 4` to `>= 4` to handle
more specific grouping.

opw-4537859

closes odoo/enterprise#80162

X-original-commit: f72b0c28b69612bce63743b97595c80a9c4949df
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Louis Gobert (loug) <loug@odoo.com>