[FIX] account_batch_payment: allow batch payments within a branch
To reproduce:
1. Create two branches within a company.
2. Set a bank account linked to the journal defined in the parent company
3. Create one payment, such as a SEPA payment or check, to be batched later,
in each branch.
4. Selecting both branches and the parent company, create a batch payment
with both payments.
5. The error "All payments in the batch must belong to the same company." is
raised.

Desired behavior:
The payments in different companies that are branches of a common head office
should not trigger the error. They should be allowed to be batched.

Cause:
The function `_check_payments_constrains()` in account_batch_payments raises
an error if more than one company_id is present in the payments being batched.

Solution:
Remove the company check in `_check_payments_constrains()`. The check on the
payments' journal_id and whether this journal_id matches the batch journal_id
is enough. Only payments from children company of the company owning the
journal of the batch payment will have a matching journal.

This way, payments from several companies, branches of the same head office,
are allowed. Payments from companies not in the same branch still raise an
error when attempting to batch them, as their journals differ.

task-4568317

closes odoo/enterprise#80285

X-original-commit: d356810
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Maira Salazar <srma@odoo.com>