[FIX] documents: sort folders in search panel
It's very difficult to find folders when they are returned in an arbitrary
order.
Furthermore, as we've removed the sequence field in 18.0, users have no way
to change the ordering of these folders.

With this change, users with many folders will at least be able to do
something until re-ordering by sequence is made possible again in 18.2

Task-4512349

closes odoo/enterprise#80386

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>