[FIX] sale_subscription: prevent payment link with renewed orders
Before this commit, it was possible to pay payment links linked to
renewed orders. It would cause issues as the renewed order would be
reopened once the transaction was set to done. Two subscription in
progress would live side by side. This commit ensure that such links
can't be created and existing links can't be used.

closes odoo/enterprise#80388

Taskid: 4607315
Signed-off-by: Thibault Francois <tfr@odoo.com>