[IMP] quality_control: improve randomly quality frequency comprehension
This commit improves the tool tip and the presentation of the quality
control random view.

When you set a quality point to "random" with a percentage of 20% and
process a transfer of 100 units, it does not mean that exactly 20% of
the transfer will be inspected. Instead, each unit has a 20% probability
of being checked.

For example, if you process ten transfers of 100 units each, you might
get 18 quality checks in one transfer, 10 in another, 25 in the next,
and so on. The key point is that, statistically, approximately 20% of
your total quantities will be inspected over multiple transfers.

opw-4527413

closes odoo/enterprise#80482

Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>