[FIX] l10n_ke_edi_oscu: Increase timeout and add logging
Recently, the Kenyan support team has been reporting a temporary
increase in the number of timeouts when sending invoices to eTIMS, as
well as associated issues (e.g. timed-out invoices turning out to be
actually registered on eTIMS' side) which warrant further investigation.

In order to reduce the number of timeouts and help diagnose (1) which
API endpoints are timing out (which can tell us where we should
implement retry mechanisms) and (2) whether the timed-out invoices are
always or sometimes registered on eTIMS' side, we:

- Increase the timeout of eTIMS calls to 2 minutes;
- Log every timeout in the server logs; and
- Log every invoice timeout in the invoice chatter.

task-4624257

closes odoo/enterprise#80726

X-original-commit: a47bcf824609c35e36a821edbbc2c655f4a3eec6
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>