[FIX] website_helpdesk_knowledge: make sure tour does not depend on demo data
Before this commit, the `helpdesk_pick_template_as_message_from_knowledge` tour
fails in no demo because the admin user is not named `Mitchell Admin`
and so the step does not find the signature field.

This commit adds the signature on the admin user to make sure the
signature will be the same with demo data and without demo data.

closes odoo/enterprise#80743

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>