[FIX] account_sepa: allow non IBAN for ISO 20022
We support sending credit transfers using ISO 20022 to accommodate non
SEPA bank transfers (any party having no IBAN number).

However when generating the XML file, we assumed that our company always
had an IBAN number and put it in an IBAN node. For non IBAN numbers
however we need to put it in another node. This commit fixes that.

opw-4585994

closes odoo/enterprise#81088

X-original-commit: e1006c9e4f70b7c9e370c1679535fb543033afec
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>