[FIX] mrp_workorder: allow printing label for finished lot before producing
Issue: you can not print label for an mo is the qty_producing is at 0.

Steps to reproduce:

- Create a storable product FP tracked by LOT.
- Create a bom for that product with 1 operation containing an
  instruction of type: "print label"
- Create and confirm an MO for 10 units and set a lot by hand (but
  ensure that the qty_producing of the MO is still at 0).
- Process the operation in the shopfloor and print label
> The printed label is empty.

Cause of the issue:

The qty of label printed depends solely on the qty_producing of the
workorder but in case this one is not set we should rather rely on the
qty to be produced.
https://github.com/odoo/enterprise/blob/2a91cb194d070879b5e1e5f5fa03178aa2dcc7ec/mrp_workorder/models/quality.py#L344-L351
https://github.com/odoo/enterprise/blob/2a91cb194d070879b5e1e5f5fa03178aa2dcc7ec/mrp_workorder/models/quality.py#L327-L332

opw-4583573

closes odoo/enterprise#80802

Signed-off-by: Steve Van Essche <svs@odoo.com>