[FIX] l10n_ar_edi: populate Private Key settings field after generation
Steps to reproduce:

1. Install l10n_ar_edi
2. Use an Argentinian company
3. Go to the "Argentinean Localization" section of the settings
4. Make sure "Private Key" and "Certificate" are empty. Save.
5. Click on "Generate Request".
6. A key has been created but looks like it's not set on the company.

Description of the issue this commit addresses:

When generating a certificate request without selecting a private key to
generate the certificate request with, a private key is generated by the system
but in the settings windows, the field remains empty.

Also, later, after retrieving the certificate from the official platform and
trying to insert it in the database, it is required to link it with the private
key that was used for its generation but since all automatically generated keys
share the same name, this can be quite hard.

Desired behavior after this commit is merged:

When using the "Generate Request" button with no private key selected, the
Private Key fields is populated with the key that was created for the
generation of the request.

Also, the names of the automatically generated keys use the current time for
differenciation purposes.

Note on the fix:

The error happened because when using the Generate Request button, the key
is created by a method called via another request so virtually, the assignment
happens in another window. To fix that issue, the key has to be created and
assigned before making any other request.

task-4393259

closes odoo/enterprise#79253

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>