[FIX] sale_subscription: compute churn amount based on previous logs
Steps to reproduce:
- Install `sale_subscription`
- Create a sub with a discount
- Confirm and invoice it
- Renew the sub without the discount
- Confirm the renewal
- Remove the discount on the original sub via the shell since it's
  locked (simulate first year discount removal)
- Cancel the renewal

Issue:
MRR log is now negative, this is due to how the way `amount_signed` is
computed during the churn event creation. Recurring monthly is now
greater because we deleted the first year discount, which results in
negative MRR.

https://github.com/odoo/enterprise/blob/5a7b55f9e059c3b54492e184d1e2429779399d08/sale_subscription/models/sale_order_log.py#L174

Solution:
Compute `amount_signed` using the sum of the previous logs.

closes odoo/enterprise#80872

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>