[FIX] sale_commission: show commission without team in report
Sales made without a team (like in POS by default) did not show
up in the commission report "details".
This commit backports the change made in later versions.

**Steps to reproduce**
- Create a commission plan for a Salesperson.
- Create an invoice from a sale in POS.
- The commission appears in the `sale.commission.report`, but
not in the details (`sale.commission.achievement.report`) if
there are other invoices for that period with a team set.

opw-4488615

Part-of: odoo/enterprise#81036
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>