[FIX] helpdesk: prevent error while uninstalling module
Currently an error occurs when the user tries to uninstall
`Sales` or `helpdesk_sale_timesheet` after creating a helpdesk ticket.

Steps to generate an error:
- Create a database without a demo and install `helpdesk_sale_timesheet`
- Create a helpdesk ticket
- Now try to uninstall sale_management (Sales)

This issue was generated because in the helpdesk ticket we have the
`use_helpdesk_sale_timesheet` field (see [1]), which is a depends field
of the  `compute_sale_line_id` method, which is the compute method
of the field sale_line_id (see [2] and [3]) in the helpdesk ticket.

In helpdesk code line [4] tries to update the value of the field
`use_helpdesk_sale_timesheet` since this field is related to the helpdesk
ticket (see [1]) value of this field also changed in the helpdesk ticket,
and the value was changed the compute method trigger because this field
depends on the compute method.

This commit will solve the above issue by executing an SQL query instead
of writing the team field so it doesn't trigger depending on that field.

[1]- https://github.com/odoo/enterprise/blob/1619b74b7895108a80d466c57972fc97923c860c/helpdesk_sale_timesheet/models/helpdesk_ticket.py#L27
[2]- https://github.com/odoo/enterprise/blob/1619b74b7895108a80d466c57972fc97923c860c/helpdesk_sale_timesheet/models/helpdesk_ticket.py#L31-L38
[3]- https://github.com/odoo/enterprise/blob/1619b74b7895108a80d466c57972fc97923c860c/helpdesk_sale_timesheet/models/helpdesk_ticket.py#L77-L78
[4]- https://github.com/odoo/enterprise/blob/1619b74b7895108a80d466c57972fc97923c860c/helpdesk/models/ir_module.py#L15

sentry-6308112739

closes odoo/enterprise#81963

X-original-commit: 7b303b0f8634da67fb08ca015fb036ec349741e1
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>