[FIX] l10n_de_pos_cert: fix payment type fiskaly
When making a payment using a payment method of type cash with a name
different than "CASH" the payment type was not being set correctly in
the fiskaly payload.

Steps to reproduce:
-------------------
* Setup fiskaly for the German POS
* Create a payment method of type cash but give it any other name than
  "CASH"
* Make a payment using that payment method
* Check the payment type on the fiskaly website dashboard
> Observation: The payment type is set to "Non cash"

Why the fix:
------------
We should base the payment type on the payment method type and not on
the payment method name.

opw-4606731

closes odoo/enterprise#80919

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>