[FIX] sale_stock_renting: account for moves in forcast availability
Issue:

The current qty forecast of rental orders does not rely on incoming and
outgoing quantities that could happen prior to the location period and
would hence modify the availability forecast.

Steps to reproduce:

- Create a storable product that can be rented put 100 units in stock.
- Create a sale order for 10 units of that product, confirm and
  confirm the delivery planned for next week.
- Create a rental order for 5 units of that product for a period
  ulterior to the delivery.
- Look at the forecast rentable quantity.
> It should be 90 but it is 100.

Cause of the issue:

The forcasted quantity of a rental order line is based on the
`qty_available` of the product:
https://github.com/odoo/enterprise/blob/01b6a70348ffa11a9ead2558571b1ccfa1dd3cbb/sale_stock_renting/models/sale_order_line.py#L115-L129
https://github.com/odoo/odoo/blob/3264e3399918b9627693e78d684ac9ac4ad3bd7d/addons/stock/models/product.py#L31-L43
However, this quantity does not take the incoming and outgoing moves
that are planned to happend prior or during the location.
In case the renting period starts at an ultarior date than today, it
should rather be based on the `virtual_available` quantity of the
product:
https://github.com/odoo/odoo/blob/3264e3399918b9627693e78d684ac9ac4ad3bd7d/addons/stock/models/product.py#L44-L55
(note that since we need to take into account each of the incoming and
outgoing moves that could happen prior to the location, the `from_date`
context key needs to be set to today).

opw-4552760

closes odoo/enterprise#81653

X-original-commit: 617243d114118bf842d815a6a89445ae3de18876
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>