[FIX] documents: prevent action not found on /odoo/documents/<access-token> click
Steps to reproduce
==================

- Install documents,project
- Go to documents
- Open any document
- Click on share
- Click on "Copy link"
- Go to a project task
- Click on a log note
- Open the full composer
- Type some text and select it
- Paste the copied url
- Post the log note
- Click on the link

=> Missing Action: the action "..." does not exist

Cause of the issue
==================

Documents [redirects] `/odoo/documents/<access-token>` to either the
backend or the portal.

There is a global click handler in the router that intercepts click on
links to update and push the router state.

Solution
========

We should not handle those links, the http redirection will take care of
it.

---

[redirects]: https://github.com/odoo/enterprise/blob/882e4453c8eadae8e39ea02314377c617d3b9e0f/documents/controllers/home.py#L24

opw-4611260

closes odoo/enterprise#81016

Signed-off-by: Julien Castiaux (juc) <juc@odoo.com>