[FIX] account_reports: fix traceback in comparison filters
Steps to reproduce:
- Go to Accounting > Configuration > Accounting Reports > Consolidated Balance > Options
- Tick "Period Comparison"
- On a consolidation period, click "Consolidated Balance"
- Click on the button "Comparison" to add a filter

Cause:
The dropdown component that is supposed to appear when clicking the button is reading value that are undefined: `controller.options.date.mode`
The options for "date" is deleted for the consolidation report:
https://github.com/odoo/enterprise/blob/ca935872fd70b9fc5b2d13359424916b3afa01eb/account_consolidation/report/trial_balance.py#L38

Solution:
We don't want to have the date filter on consolidation balances but we want to keep the consolidation filter.
So we check if `controller.options.date` exist before reading `mode`.

opw-4546574

closes odoo/enterprise#80492

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>