[FIX] l10n_es_real_estates, l10n_es_reports: fixed boe key map of real estates invoices
Steps to reproduce the bug:
- Install l10n_es_real_estate module
- Create a customer invoice on the accounting app
  - Invoice's AEAT data should be real estate type for mod347 doc
- Generate the BOE of tax report document of mod 347

Traceback is thrown while generating the boe of the mod347 document,
the traceback is for an issue related to the param of the operation
key and that was because the function _call_on_partner_sublines is run
for each real estate invoice and it has a callback to be executed on
each of them.

The callback function is _write_type2_partner_record which should have
the report option as param, but it wasn't sent that made a traceback
for the params.

After fixing that, another traceback was thrown because the xmlid of
the real estate invoices of both sold and bought are not in the
invoice types map of _write_type2_partner_record function.

opw-4589314

closes odoo/enterprise#82121

X-original-commit: 488058198db1638a5bf46f40821c85d4334f7b63
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Hussein Yasser Hussein Ebrahim (hyhe) <hyhe@odoo.com>