[FIX] test_web_studio: usage of docstring in 3.13
Since 3.13, leading whitespace is automatically stripped in a
docstring (python/cpython#81283).

The `info` field of an ir.model is a copy of the docstring, so running
:TestStudioExports.test_export_customizations on 3.13 fails because
the documents are different (the content of the info field is indented
and has a leading space matching `Model`'s docstring).

Instead of duplicating the info field, just inject it in the document
we're creating from the record, that way it doesn't break if someone
decides to improve the docstring either.

closes odoo/enterprise#81045

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>