[FIX] helpdesk: Use local tz for dashboard and actions
eLearning uses server date when processing dashboard values, which causes
a mismatch from the perspective of the user. This commit changes the
source of "today" to use the user's timezone on the dashboard and
actions, while keeping compute functions on the server date as these
values should not change based on the timezone of the observer.

opw-4411615

closes odoo/enterprise#80490

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>