[FIX] web_studio: blacklist knowledge.article from many2manys
Steps to reproduce
==================

- Install knowledge,contacts,web_studio
- Open a contact record
- Open studio
- Add a many2many pointing to a knowledge article
- Save and exit studio
- Link an article and open it

=> this.env.ensureArticleName is not a function

Cause of the issue
==================

The knowledge article view uses a js_class and contains widgets that
except the custom subenv from the controller.
As the js_class isn't used in x2many views, the view crashes.

Solution
========

As we cannot change the knowledge views in stable to be usable without
the modified context, we blacklist that model when searching for relations.

opw-4513295

closes odoo/enterprise#81068

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>