[FIX] industry_fsm_stock: fetch the right warehouse in test
Before this commit, if the `property_warehouse_id` field is set on the
current user inside the test, the expected warehouse could not be found
with the search made on the test (if more than one warehouse exists on
the second company used in the test).

This commit makes sure the test expects the right warehouse by calling
the `_get_default_warehouse_id` defined in `res.users` model to assert
the stock move is done in the default warehouse of the current user.

closes odoo/enterprise#81330

X-original-commit: a8dd40b244ca9c2fac296d6afb184a78a15eda7e
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>