[FIX] documents: restore colorless tags on kanban cards
Even though it's been commonplace in Odoo for a while,
tags (or more precisely facets) with index-0 color where still shown
on Documents kanban cards.

This should remain the case functionally as users are not expected to
go and change the color of all their tags, nor can we decide for them
at upgrade (and it'd be too late for many as well).

Task-4645015

Part-of: odoo/enterprise#76201
Signed-off-by: Patrick Hoste (pko) <pko@odoo.com>