[FIX] stock_barcode: don't split done move
Issue
=====
When validating an uncompleted return in Barcode, an error is raised.

How to reproduce
================
- Open Barcode app > Operations > Receipts and create a new receipt;
- Scan at least two times the same product's barcode;
- Validate.
- Re-open the same receipt and click on "Return Products";
- Scan one product then validate -> The return is correctly validated
  but an "Invalid Operation" error is raised.

Cause of the issue
==================
The return's move is done but `_split` is called and calling this method
on a done move is forbidden.

How to fix
==========
In the `split_uncompleted_moves` move's method, done and cancel move are
skipped.
Also, the JS method who call `post_barcode_process` will now doesn't
call it if operation is done or cancelled or if there is no moves (to
avoid to do useless RPC.)

OPW-4535205
OPW-4535205

closes odoo/enterprise#81312

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>