[FIX] l10n_co_dian: handle odoo and DIAN being out of sync
it could happen that odoo does not receive DIAN's response, because of internet loss while the document is being sent to dian for example.
The move form will then display the edi state as being rejected while in reality DIAN has successfully processed the move.
If the user where to send the move to DIAN again, they would get an error message saying the move has already been processed.

the fix: when the move is sent to DIAN again after getting out of sync, DIAN will send the error explained above which contains the identifier of the move. We now save this identifier and use it in the automatic GetStatus call to get odoo and DIAN back in sync.

closes odoo/enterprise#81615

Ticket: 4569829
X-original-commit: 815ae71b97a189746772c7b3994bdd92859e062e
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>
Signed-off-by: Allesio Pellicciotta (peal) <peal@odoo.com>