[FIX] knowledge: load PermissionPanel only once
__Current behavior before commit:__
When clicking on the share button of a knowledge article, the
`/knowledge/get_article_permission_panel_data` call is made twice.

This is because the call to `loadPanel` is made inside the `onWillStart`
as well as in the `useEffect`.

The call made in the `onWillStart` was already removed in [this PR][1].
However it was added back in the [saas-17.1 forward-port][2] probably by
mistake.

__Description of the fix:__
Removed the call to `loadPanel` in the `onWillStart`.

opw-4603551

[1]: https://github.com/odoo/enterprise/pull/58825
[2]: https://github.com/odoo/enterprise/pull/59561

closes odoo/enterprise#81477

X-original-commit: 8ca9f32c733f698ed2052d1c6b89f6dabbef7dc2
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>
Signed-off-by: Julien Launois (jula) <jula@odoo.com>