[FIX] l10n_es_reports: VAT Record Books export: properly handle entries with multiple similar tax lines
To reproduce the issue:

1) Install l10n_es_reports
2) Configure the taxes WHI 15% and IVA 21% purchase taxes to include analytic (checkbox on the form view)
3) Create an invoice with two lines of 100€, so that
    - both lines use both IVA21% and WHI 1% taxes
    - one of those lines defines an anlaytic distribution while the other doesn't
4) Open the generic tax report, and generate the VAT Record Books XLSX export

===> In that file, the amount of withholding is 15€. It should be 30€.

The same issue can be reproduced for surcharge taxes, in the same way.

This is due to the fact different analytic distributions between base lines will trigger the creation of different tax lines for the same tax (one for each distribution). The code handling the export "saw" both lines, but ovewrote the total instead of summing all the values.

opw-4471253

closes odoo/enterprise#81602

X-original-commit: 291bb055e0ecf99c276fed9ef3d53f4754539707
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>