[FIX] helpdesk_stock: avoid wizard if no picking found
To reproduce the issue:
1. Confirm a SO with partner P
2. Create/Edit an helpdesk team
   - Returns: True
3. Create a ticket for P
4. Open its form view
5. Click on return

Error: a traceback appears "ValueError: Expected singleton: stock.
picking()"

The return button loads the wizard `stock.return.picking`. In this
model, we try to find a related done picking:
https://github.com/odoo/enterprise/blob/7215e4abd904ebdff3bd85c2c2f3002b1f5593b1/helpdesk_stock/wizard/stock_picking_return.py#L24-L31
Which, in the above case, does not exist. Later, we call a method on
that picking:
https://github.com/odoo/odoo/blob/8f527e93de2d64111904f1fcc5813e8b8ea0e4ba/addons/stock/wizard/stock_picking_return.py#L100-L104
Where `_can_return` is a `ensure_one` method, hence the error.

The return button is only displayed in some conditions:
https://github.com/odoo/enterprise/blob/3d8bb40fcad00068e12f21b623390c959d3fd19c/helpdesk_stock/views/helpdesk_ticket_views.xml#L35
Here is the problem: `has_partner_picking` can be `True` if we find
some related SOL, which is useless, the only thing that matters are
the related pickings.

OPW-4591280

closes odoo/enterprise#81378

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>