[FIX] stock_barcode: product from parent company
Issue
=====
When logged in a branch company, it is not possible, in the Barcode app
inventory adjustment, to select a product with its `company_id` set on
the parent company.

How to reproduce
================
1. Create a branch company;
2. Create a product and set its company field on parent company;
3. Select the branch company as the current user company;
4. Go to Barcode > Inventory Count > Add Product;
5. In the product field, search the parent company's product
   -> Nothing found.

Cause of the issue
==================
There was a domain on the `product_id` field who restrain products only
to the current company (both for quants and stock move line.)

Solution
========
Adapt the product's domain to include parent company.

OPW-4603831

closes odoo/enterprise#81453

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>