[FIX] l10n_mx_edi_pos: recompute order amounts when atempting to pay
When a quotation is loaded into the POS, a down payment is made, and
then the order is settled, if the order has a discount that causes the
total amount to become negative, the system prevents payment and raises
an error. However, if the discount line is removed, the payment attempt
also raises the same error.

This happens because, even though the `pos.order.line` was removed, the
values of the `pos.order` were not updated. Since it is required to
validate the amount to be paid in the inherited `pay` method, the error
occurs.

This commit fixes the issue by recomputing the amounts before the
validations that require them.

closes odoo/enterprise#81444

Signed-off-by: Vlad Stroia (vlst) <vlst@odoo.com>