[IMP] account_followup: followup report generated when printing
Before, when sending a reminder for the followup via the send and print
wizard, the customer statement, or the followup report, were only
included if you checked email, but if only print was selected, the
report was missing in the PDF.

Now, when selecting print, email, or both, you get the report.

Additionally, this fixes an issue where disabling the "Attach Invoice"
option would remove all attachments, including the follow-up report,
when it should have only excluded invoice attachments. The follow-up
report is now always included as expected.

opw-4570715

closes odoo/enterprise#81463

Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>