[FIX] quality_mrp_workorder_worksheet, mrp_workorder: validate worksheet QC without wizard
Steps to reproduce:

- Create a BOM with an operation and an instruction of type worksheet.
- Create and confirm an MO for 1 unit.
- In teh shopfloor process the instruction and save.
> The QC is not marked as completed hence, the button Mark as done is
> not present.

Cause of the issue:

Since commit 60c02ac4f7d5626a0df4fdd81a02bb540abfdb59 it is required
for the context to contain a quality_wizzard_id to trigger a validation
of the worksheet instruction:
https://github.com/odoo/enterprise/pull/75711/commits/60c02ac4f7d5626a0df4fdd81a02bb540abfdb59#diff-9c37a7c8d7c9040a46168f49e4f144269a2b95b644fef933e29bdd6c26370776R17-R33
This is a backport of a change present in 17.4. However, in 17.4,
clicking in the instruction trigger a call of the
`action_open_quality_check_wizard` action:
https://github.com/odoo/enterprise/blob/8f30a651aefd13293eb0cdeaf2fb73d3544d8552/mrp_workorder/static/src/mrp_display/mrp_display_record.js#L315-L319
and hence adds the wizzard_id in the context.
However, in 17.0, we trigger a call of the `action_fill_sheet`
https://github.com/odoo/enterprise/blob/c466a605ac0384ad6ca1602341eb44a249a1cc2d/mrp_workorder/static/src/mrp_display/mrp_display_record.js#L311-L315
which does not.

Fix we packport the same fix as in 17.4: e08e6f07768134ed181301f72f8a45b867897b77

opw-4640022
opw-4640058

closes odoo/enterprise#81518

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>