[FIX] documents_hr: display error if hr workspace isn't set
Issue
When using the `Send HR Documents Access Link` action, Odoo always assumes that
the company has set a `documents_hr_folder` already. However, this can easily not
be the case if the customer has never activated `documents_hr_setting` and
chosen a Workspace. If this is the case, we'll encounter an error when creating
the `document.share` record, since `folder_id` is required.

Solution:
Display an error directing the customer to activate an HR Workspace folder
so they can properly use the feature.

opw-4559772

closes odoo/enterprise#82563

X-original-commit: 0de52a63dafe21c2efa81b4547c26ef7626f809b
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>