[FIX] account_bank_statement_import_ofx: decoding error
When importing an ofx bank statement, we try to parse
the raw attachment value, in case of UnicodeDecodeError,
we try to get an encoding declaration in the file, and fallback
on 'utf-8'. But we can still face a file that is not well
encoded and we end up with a traceback.

With this commit, we display a UserError instead of the ugly traceback.

opw-4526766

closes odoo/enterprise#82458

X-original-commit: 6ead5525a9dceb6f53d4f06f6fe98228c967cb8c
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>