[FIX] account_bank_statement_import_csv: Auto-reconcile on CSV import
Problem
---------
When drag-and-dropping a CSV in the bank journal, the statements lines
don't get reconciled directly while they should (if a reco model allows
it). Users have to wait for the auto-reconcile cron to run.

Solution
---------
After the statement and its lines have been created through import, we
trigger the cron computation.

opw-4545446

closes odoo/enterprise#81958

X-original-commit: 77ba255fed3157939b67aa5c0f4f7782c91c8857
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>