[FIX] sign: test_sign_flow_tour
In this commit, we fix the latest step of tour. We can't click on
primary or secondary button ... make no sense.
As the primary button download the pdf file and there is no test to
ensure the file is well downloaded, it make more sense to click on
close the modal.
With the fix in https://github.com/odoo/odoo/pull/202089 and without
the correction, the tour was timeout exceeded.

closes odoo/enterprise#81596

Related: odoo/odoo#202093
Signed-off-by: Bastien Pierre (ipb) <ipb@odoo.com>