[FIX] l10n_be_coda: date not in index range
Description of the issue this commit addresses:

A traceback is raised when importing some .cod files for bank transactions
because the server tries to create a date element with a string that doesn't
match the expected format.

Steps to reproduce:

1. Install l10n_be_coda
2. Go to the accounting dashboard and import the file "02-minimal-test.cod"
   from the ticket.
3. A traceback is raised

Desired behavior after this commit is merged:

The file import doesn't raise a traceback.

Note on the fix:

The issue happened because in `_get_coda_file_statement()`, the date is
expected to be on the index range 115-120 but the the decoding of the file in
`_parse_bank_statement_file()` transformed a "€" in "Ôé¼" which moved the date
to indexes 117-122. This wrong translation of encoding happened because cp850
was used - which doesn't support currency symbols - instead of utf_8 because
the latter was stopped by a non matching regex that was missing the currency
symbols.

This commits includes currency symbols in the validating regex so that the most
appropriate encoding is validated as often as possible.

opw-4546281

closes odoo/enterprise#81329

Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>