[FIX] l10n_es_reports: MOD347 operator count with journal entry
Steps to reproduce:
- With an ES company setup
- Create and confirm invoice with:
  - Spanish partner
  - Amount > 3005.06 (mod347 threshold)
  - Type for mod 347: Regular operation
- Create and confirm a jounral entry with:
  - Payable account, debit 4000
  - Receivable account, credit 4000
  - Type for mod 347: Regular operation
- Check Mod 347 Tax Report

Issue: 'Total number of persons and entities' shows 0

This occurs because some lines of mod 347 report need to be grouped by partner,
only keeping the partners whose balance for the line is above 3005.06€,
so we first get all the partners that match the domain but don't reach the threshold.
We exclude these partners with a 'NOT IN' clause. However, when the partner is not set, a
NULL values is retrieved causing the clause to be evaluated NULL instead of False
and the total count will be 0

opw-4544950

closes odoo/enterprise#81687

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>