[FIX] planning_contract: gantt view default start and end datetime fixed for when scale is set to day
Prior to this commit, 2 issues were observed for flexible resources in the gantt view.
First when the scale of the gantt was set to 'day' (meaning, each cell was
representing an hour) and the interval was also on 1 day (ex: 18th March)
flexible resources without contracts were not able to fetch their working periods
leading to a gray cells for the whole period expect for the 1st hour of the day.

Secondly, when the scale of the gantt is set to 'day', whilst having a larger
interval over multiple days (ex: 15th - 22th), the default_end datetime was
missing 1 day in the interval, leaving grey cells at the end of the planning
view even though flexible resources should have these cells white.

Step to reproduce the first issue:
1. Install Planning, HR_Contract
2. Create a resource with flexible 40h calendar, without a contract.
3. Open Planning > Set the scale to hour granularity with a single day as interval (ex: 16th March)
4. The whole day should be greyed out expect for the 1st hour. (missing working periods)
-> Should be white as the resource is flexible and should be able to work the whole day.

Step to reproduce second issue:
1. Install Planning, HR_Contract
2. Create a resource with flexible 40h calendar, without a contract.
3. Open Planning > Set the scale to hour granularity with the whole week as interval (ex: 16th - 22th March)
4. When scrolling to the end of the week, the last day should be greyed out. (missing 1 day)
-> Should be white as the resource is flexible and should be able to work the whole day.

This commit replaces the argument passed to the default_start/end to be
the global_start/end of the gantt view.

related bugfix task-id: 4492625

closes odoo/enterprise#79532

Related: odoo/odoo#198032
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>