[FIX] whatsapp: avoid conflicts on base data creation
It is possible for a user to sync an existing whatsapp account
and only after install some whatsapp bridge module.

In that case the user will get an error as the template is created
on the existing account with an existing name.

We now prevent using the default value if a template with the same name
already exists on it.

task-4369979

closes odoo/enterprise#79786

closes odoo/enterprise#82337

X-original-commit: 833972837002663f7609f758ab60983e17063fd6
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>