[FIX] appointment: fix undeterministic test
This commit is meant to fix a test in appointment failing randomly.
The issue likely came from the fact that records were altered several
times on the same model within the same test, with the gantt view being
destroyed and re-created each time.

To avoid this issue altogether, the test has been split for each case.

closes odoo/enterprise#81735

Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>