[FIX] l10n_nl_reports_vat_pay_wizard: no belastingdienst account
When updating the module, the bank data is reloaded. When the account is already trusted (`allow_out_payment == True`),
the write fails due to checks to prevent fraud. To fix this, we wrap the data in a `noupdate` to prevent further
writes from happening. If ever the account number changes, a new record will be
created instead of updating the existing one.

Since this account is the default one stipulated on the official belastingdienst site, it can be trusted by default.

https://www.belastingdienst.nl/wps/wcm/connect/bldcontenten/belastingdienst/business/payroll_taxes/you_are_not_established_in_the_netherlands_are_you_required_to_withhold_payroll_taxes/when_you_are_going_to_withhold_payroll_taxes/filing_payroll_tax_returns_and_paying_payroll_tax/payment

closes odoo/enterprise#81740

Similar: d4595b856045fe9d35a4ac0f28faf12f0d19cd88
Signed-off-by: John Laterre (jol) <jol@odoo.com>