[FIX] web_studio: new kanban: overflowing elements should be reachable
In studio, editing a kanban, activate show invisible elements if needed
Before this commit, when the footer had many items, those were not wrapped
so, they overflew under the next card and were unreachable and unclickable.

After this commit the overflowing elements are reachable and clickable, but still outside of the kanban box.
They just have a z-index high enough to be visible and reachable.

The motivation behind this choice is that we cannot really apply targetted visual rules as those would betray
the actual visual on the view in normal mode.

task-4609581

closes odoo/enterprise#81820

Signed-off-by: Bruno Boi (boi) <boi@odoo.com>