[FIX] account_accountant_batch_payment: test reset of batch payment
Test that resetting a statement line that was merged with a batch
payment also resets the payments, especially for payments without
an outstanding account.
It wasn't the case because there is no real link between the batch
payment and the statement line, and neither between the statement line
and the payment itself.

The fix is in community.

closes odoo/enterprise#81839

Related: odoo/odoo#202692
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>