[FIX] website_sale_renting: display correct amount of days when in cart
Versions
--------
- 17.0+

Steps
-----
1. Have a rental product with a minimum rental duration of 1 day;
2. open its eCommerce page;
3. select two days in the rental period datepicker;
4. go to cart;
5. go back to the product page and refresh.

Issue
-----
The product is displayed with a rental period of 3 days.

Cause
-----
Commit 0573ad5b7007 attempted to add a timezone correction when
retrieving DateTime values from the server. When duration isn't with
hours, it gets the `startOf` of the start datetime, and the `endOf` of
the end datetime as a means of converting datetimes to dates.

Issue is that if start date is e.g. 2025-03-23 23:00:00 and end date is
2025-03-25 22:59:59 due to timezone correction, the `startOf` of the
start date is 2025-03-23 and the `endOf` of the end date is 2025-03-25
due to being adjusted -1 hours from UTC.

This results in an apparent time delta of 3 days, while the actual
delta should be 2 days.

Solution
--------
Leave the datetime values as is, as there is other logic in place that
correctly converts them to date values on the frontend.

opw-4583312

closes odoo/enterprise#81849

Signed-off-by: Valentin Chevalier <vcr@odoo.com>