[FIX] sale_purchase_stock_inter_company_rules: fix dropship test
The test requires `stock_dropshipping` to be installed to have sense,
but the condition set to avoid the test running without it was
insufficient.
Since it was only checking on a string in the database, it wouldn't
raise a ValueError, but just not find any matching record, thus running
the test even without the module installed.

closes odoo/enterprise#82016

X-original-commit: 94f4e327748ae5509a0147f590678824311ed0dd
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>