[FIX] hr_attendance_gantt: correctly compute max_value in progressbar
Before this commit, when the employee has flexible working schedule, the
company full time field is used to compute the max_value for the
flexible employees. The problem is the number of days taken into account
is one day less than the period displayed in the gantt view, which
means if the employee is supposed to work 40 hours for one week, the
progress displayed in the gantt view.

This commit adds one day to include the end date in the computation and
adds a test to make sure the progress bar of employee fetched in gantt
view of attendance returns the expected values without any issue.

opw-4492625

closes odoo/enterprise#82068

Related: odoo/odoo#203253
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>