[FIX] account_reports: Fixed reports scroll position on navigation
Since 17.0, and before this commit, when the user navigates from an account report to another page
and then comes back to the report through the breadcrumb, the user is brought to the top of the report.
In 16.0, the user would be brought to the same position they were looking at.
This commit brings this behavior back.

task-4578468

closes odoo/enterprise#82465

X-original-commit: 1ae81683b47a0cc8e27663cd34ef415ee261aaf0
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>