[FIX] spreadsheet_sale_management: useless read
Users don't need to be able to read all spreadsheet.
It hurts performance.

Note that it's not a security vulnerability to be able to read the
spreadsheet. The user cannot read the sale order (lines) as all other
security rules apply. The only thing the user can read in the spreadsheet
are formulas (which he is able to read in the spreadsheet template anyway)

closes odoo/enterprise#81728

Task: 4656165
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>