[FIX] sale_purchase_inter_company_rules: don't create rentals from POs
Versions
--------
- 18.0+

Steps
-----
1. Have purchase & sale_stock_renting installed;
2. create a 2nd company;
3. enable inter-company transactions for Sale/Purchase Orders for both;
4. enable Warehouse storage locations;
5. enable multi-step routes;
5. enable dropshipping;
6. enable rental transfers;
7. create a dropshipping route assigned to the 1st company;
8. create a purchasable, rental product, tracked by quantity;
9. add dropship route to its operations;
10. add 2nd company as vendor;
11. create a rental order with the product;
12. confirm the rental order;
13. go to the generated purchase order via the smart button;
14. confirm the purchase order;
15. switch to the 2nd company;
16. try to open the generated sale order.

Issue
-----
> TypeError:
> '>' not supported between instances of 'datetime.datetime' and 'bool'

Cause
-----
On confirming the purchase order, the sale order it generates gets
created as a rental order, due to the `in_rental_app` context value
being set.

When trying to open the rental order, it attempts to compute the
quantity available for the tracked rental products during the rental
period.

The rental order was generated based on purchase order data, which lacks
any kind of rental information, hence the rental order doesn't have a
rental period set.

The lack of rental dates is what causes the error when trying to
partition the lines based on rental period.

Solution
--------
When creating a sale order via a purchase order, pass the
`in_rental_app=False` context value to ensure it won't be created as a
rental order.

opw-4583210

closes odoo/enterprise#82220

Signed-off-by: Levi Siuzdak <sile@odoo.com>