[FIX] sale_subscription: fix cron trigger for subscription
Issue:
========

There is an issue during the generation of invoices from subscription
that causes the is_invoice_cron field on the sale order to remain `True`.
This prevents the processing of transactions and, as a result, does not
mark the invoices as paid. Furthermore, this issue blocks the generation
of subsequent invoices.

To reproduce:
=============

1. Create a subscription that does not require initial client payment,
(can generate an invoice on the next invoice date).
2. Create 30 other subscriptions (equal to the batch size value)
that cannot generate invoice until the client has made a payment.
3. Run the scheduled action “Sale Subscription: generate recurring invoices and payments.”
4. You will notice that the first Sale Order remains to “is_invoice_cron” as True

Problem:
========
the condition to trigger the cron to invoice subscriptions is based on the
number of subscriptions that can be invoiced in a batch. But we are taking
all the subscriptions while it should be only the subscriptions that can be
invoiced.

Solution:
=========
filter the subscriptions that can be invoiced before checking the condition
to trigger the cron to invoice subscriptions.

opw-4559250

closes odoo/enterprise#82340

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>