[FIX] website_sale_renting: display correct amount of days when in cart
Versions
--------
- 17.0+

Steps
-----
1. Have a rental product with a minimum rental duration of 1 day;
2. open its eCommerce page;
3. select two days in the rental period datepicker;
4. go to cart;
5. go back to the product page and refresh.

Issue
-----
The product is displayed with a rental period of 3 days.

Cause
-----
Commit 0573ad5 attempted to add a timezone correction when
retrieving DateTime values from the server. When duration isn't with
hours, it gets the `startOf` of the start datetime, and the `endOf` of
the end datetime as a means of converting datetimes to dates.

This only works correctly if the parses datetime value was serialized
with a timezone value. This was already done for the default start dates
passed to `combination_info`, but not when retrieving date fields from a
current sale order.

Solution
--------
Use the `_get_localized_renting_dates` method in the template to get the
relevant date values with the correct timezone offset.

opw-4583312

closes odoo/enterprise#82704

X-original-commit: caf059ddc08a9a59ba4277d3ed48d9a85bf9af5a
Signed-off-by: Valentin Chevalier <vcr@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>