[IMP] l10n_lu_reports: 2025 coa and vat report update
Description of the issue this commit addresses:

In january 2025, updates were made to some sections of the LU tax report and
chart of accounts. Therefore, the version we use is out of date.

Desired behavior after this commit is merged:

The COA and VAT reports use the new values updated in 2025.

task-4587067

closes odoo/enterprise#82627

X-original-commit: https://github.com/odoo/enterprise/commit/e73c06b7eb10aff95c05468710c2f9a44f0892b2
Related: odoo/odoo#204232
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>